Move state into hooks/Provider
